[Refactor] DB 반영, 국어형식 터지는 상태

+++ semomun/Solving/SaveSectionUsecase.swift
@@ -11,8 +11,8 @@ import CoreData
 class SaveSectionUsecase {
     let section: Section_Core
     var vids: [Int] = []
-    var perfactScore: Int = 0
-    var totalScore: Int = 0
+    var perfactScore: Double = 0
+    var totalScore: Double = 0
     var wrongProblems: [String] = []
     var submissions: [Submission] = []
     
@@ -71,11 +71,11 @@ class SaveSectionUsecase {
             
             if let correct = submission.correct,
                let pName = problem.pName {
-                self.perfactScore += Int(problem.point)
+                self.perfactScore += problem.point
                 if correct == 0 {
                     self.wrongProblems.append(pName)
                 } else {
-                    self.totalScore += Int(problem.point)
+                    self.totalScore += problem.point
                 }
             } else {
 //                self.totalScore += Int(problem.point)
